feat : 키워드 알림- 키워드 조회 API에서 전체 키워드 개수 출력 하도록 수정

+++ src/main/java/shop/makaroni/bunjang/src/service/SettingService.java
@@ -47,7 +47,10 @@ public class SettingService {
         settingDao.patchNotification(userIdx, res);
     }
 
-    public Long postAddress(Long userIdx, Address req) {
+    public Long postAddress(Long userIdx, Address req) throws BaseException {
+        if(settingDao.getUserAddress(userIdx).size() >= 10){
+            throw new BaseException(SETTING_ADDRESS_CAPACITY);
+        }
         Long addrIdx = settingDao.postAddress(userIdx, req);
         settingDao.deleteAddrDefault(userIdx, addrIdx);
         return addrIdx;
